Split the CVE-2022-30333 unrar module into two different modules with a shared mixin to generate the file
